@@ -9,26 +9,29 @@
     "build": "cross-env NODE_OPTIONS=--max-old-space-size=4096 vite build",
     "serve": "vite preview",
     "test": "jest src",
+    "lint": "eslint src --ext .js,.jsx,.ts,.tsx",
     "typesync": "npx typesync"
   },
   "dependencies": {
     "@duckdb/duckdb-wasm": "^1.20.0",
-    "@mockingjay-io/react-diff-viewer": "^1.0.0",
-    "@mui/material": "^5.6.4",
-    "@primer/octicons-react": "^13.0.0",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/material": "^5.11.1",
+    "@primer/octicons-react": "^17.10.0",
     "@react-icons/all-files": "^4.1.0",
     "bootstrap": "^4.6.0",
+    "dayjs": "^1.11.7",
     "http-proxy-middleware": "^1.1.1",
     "lodash": "^4.17.21",
-    "moment": "^2.29.2",
     "query-string": "^7.1.1",
-    "react": "^17.0.0",
+    "react": "^18.2.0",
     "react-bootstrap": "^1.5.2",
-    "react-dom": "^17.0.0",
+    "react-diff-viewer-continued": "^3.2.3",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.3.1",
     "react-ipynb-renderer": "^0.3.0",
     "react-markdown": "^8.0.3",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.5.0",
     "react-syntax-highlighter": "^15.5.0",
     "remark-gfm": "^3.0.1",
     "remark-html": "^15.0.1",
@@ -39,25 +42,29 @@
     "@types/bootstrap": "^4.6.2",
     "@types/jest": "^29.0.3",
     "@types/lodash": "^4.14.185",
-    "@types/react": "^17.0.50",
-    "@types/react-dom": "^17.0.17",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
     "@types/react-router-dom": "^5.3.3",
     "@types/react-syntax-highlighter": "^15.5.5",
     "@types/react-test-renderer": "^17.0.2",
     "@types/validator": "^13.7.7",
     "@typescript-eslint/eslint-plugin": "^5.38.0",
     "@typescript-eslint/parser": "^5.38.0",
-    "@vitejs/plugin-react": "^2.2.0",
+    "@vitejs/plugin-react": "^3.0.0",
+    "@vitejs/plugin-react-swc": "^3.0.1",
     "babel-eslint": "^10.1.0",
+    "browserslist": "^4.21.4",
     "cross-env": "^7.0.3",
+    "eslint-plugin-compat": "^4.0.2",
     "eslint-plugin-react": "^7.23.2",
     "jest": "^29.0.3",
-    "react-test-renderer": "^17.0.2",
+    "react-test-renderer": "^18.2.0",
+    "rollup-plugin-analyzer": "^4.0.0",
     "ts-jest": "^29.0.2",
     "ts-node": "^10.9.1",
     "typescript": "^4.8.3",
     "typesync": "^0.9.2",
-    "vite": "^3.2.3",
+    "vite": "^4.0.2",
     "vite-plugin-eslint": "^1.8.1"
   }
 }
